pyproject.toml
[tool.poetry]
name = "flynapse-docs-scripts"
version = "0.1.0"
description = "Utility scripts for documentation (docs/script-tools)."
authors = ["Flynapse Engineering Team <engineering@flynapse.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
requests = "*"
openai = "*"
python-dotenv = "*"
pydub = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
coverage = "*"
black = "*"
flake8 = "*"
isort = "*"
mypy = "*"
pre-commit = "*"
pytest-asyncio = "*"
ruff = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\\.pyi?$'
extend-exclude = '''
/( 
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["docs"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short"
testpaths = ["tests"]


